Add files via upload
I have opted to separate the mySQL connection and setup from the main file. This is now part of the python_SQL_functions.py file with the other helper functions.
I then tested the functionality of the new structure and it performs exactly as before.
There are also some functions to create some data visualisations towards the latter section of the code.
Next step will be to extract and parse the pertinent information from the result of SQL queries.